feat: create flow of authentication

+++ server/src/routes/memories.ts
@@ -1,10 +1,18 @@
 import { FastifyInstance } from 'fastify'
 import { z } from 'zod'
+
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        user_id: request.user.sub,
+      },
       orderBy: {
         created_at: 'asc',
       },
@@ -21,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -33,6 +41,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.is_public && memory.user_id !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return {
       memory,
@@ -53,7 +65,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         cover_url,
         is_public,
-        user_id: 'bc2f4de6-b281-42e5-8e20-6ad9a40d537f',
+        user_id: request.user.sub,
       },
     })
 
@@ -75,6 +87,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, cover_url, is_public } = bodySchema.parse(request.body)
 
+    const memory = await prisma.memory.findUnique({
+      where: {
+        id,
+      },
+    })
+
+    if (!memory) {
+      return reply.status(400).send({
+        message: 'Memory not found.',
+      })
+    }
+
+    if (memory.user_id !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     await prisma.memory.update({
       where: {
         id,
@@ -95,6 +123,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUnique({
+      where: {
+        id,
+      },
+    })
+
+    if (!memory) {
+      return reply.status(400).send({
+        message: 'Memory not found.',
+      })
+    }
+
+    if (memory.user_id !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
